Edited Exception log class, added movement crud operations.

+++ Core/IEntityRepository.cs
@@ -11,10 +11,10 @@ namespace Core
     public interface IEntityRepository<T> where T : class, IEntity, new()
     {
         Task<int> AddAsync(T entity, string procedure);
-        Task<int> DeleteAsync(long id, string procedure);
+        Task<int> DeleteAsync(int id, string procedure);
         Task<int> UpdateAsync(T entity, string procedure);
         Task<IEnumerable<T>> GetAllAsync(string procedure);
-        Task<T> GetById(long id, string procedure);
+        Task<T> GetById(int id, string procedure);
 
     }
 }
